Version 1.8 mit Doku aktualisert
git-svn-id: svn://localhost/Transferprojekt@126 d8cf526a-ed47-ab48-8470-8a9acf7b5d78

+++ csv2siard/c2sconvert.php
@@ -320,7 +320,7 @@ function writeSchemaContent($siardhandle, &$table){
 				case "FLOAT":
 				case "REAL":
 				case "DOUBLE":
-					$xstype = 'double'; break;
+					$xstype = 'float'; break;
 				case "NUMERIC":
 				case "DECIMAL":
 					$xstype = 'decimal'; break;
